Split `models` file into directory
